% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProAE.R
\docType{data}
\name{PROCTCAE_table}
\alias{PROCTCAE_table}
\title{PRO-CTCAE variable / label crosswalk}
\format{
A data frame with 124 rows and 2 variables
}
\description{
A crosswalk / look-up table of expected variable names for associated PRO-CTCAE symptom items.
}
\details{
\itemize{
  \item name. Expected variable name - item number/letter corresponds to the NCI-PRO-CTCAE (English) Item Library Version 1.0
  \item short_label. Item label including the item symptom group and frequency, severity, interference, or presence component
}
}
\references{
\url{https://healthcaredelivery.cancer.gov/pro-ctcae/instrument-pro.html}
}
\author{
Blake Langlais
}
\keyword{data}
