% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProAE.R
\docType{data}
\name{tox_chronic}
\alias{tox_chronic}
\title{PRO-CTCAE data reflecting chronic drug toxicity}
\format{
A data frame with 1400 rows and 5 variables
}
\description{
Simulated example data where the drug group experiences chronic toxicity over the course of treatment.
}
\details{
\itemize{
  \item id. Subject identifier
  \item Cycle. Time variable denoting visits/cycles (1-10)
  \item arm. Treatment groups include drug and placebo
  \item PROCTCAE_9A_SCL. PRO-CTCAE frequency item for nausea
  \item PROCTCAE_9B_SCL. PRO-CTCAE severity item for nausea
  \item PROCTCAE_9_COMP. PRO-CTCAE composite score for nausea
 }
}
\author{
Blake Langlais
}
\keyword{data}
