% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProAE.R
\docType{data}
\name{tox_late}
\alias{tox_late}
\title{PRO-CTCAE data reflecting late incipient drug toxicity}
\format{
A data frame with 1400 rows and 5 variables
}
\description{
Simulated example data where the drug group experiences late incipient toxicity towards the end of the treatment period.
}
\details{
\itemize{
  \item id. Subject identifier
  \item Cycle. Time variable denoting visits/cycles (1-10)
  \item arm. Treatment groups include drug and placebo
  \item PROCTCAE_78A_SC. PRO-CTCAE frequency item for nosebleeds
  \item PROCTCAE_78B_SC. PRO-CTCAE severity item for nosebleeds
}
}
\author{
Blake Langlais
}
\keyword{data}
