% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxFigures.R
\name{toxFigures}
\alias{toxFigures}
\title{Create PRO-CTCAE severity frequency distribution figures for individual
survey items and composite scores}
\usage{
toxFigures(
  dsn,
  id_var,
  cycle_var,
  baseline_val,
  arm_var = NA,
  plot_limit = NA,
  colors = 1,
  bar_label = 0,
  cycle_label = FALSE,
  cycle_vals = NA,
  cycle_labs = NA,
  summary_only = FALSE,
  summary_highlight = FALSE,
  cycles_only = TRUE,
  x_lab_angle = 0,
  x_lab_vjust = 1,
  x_lab_hjust = 0,
  x_label = "Randomized Treatment Assignment",
  y_label = "Percent of Total Frequency",
  footnote_break = FALSE,
  suppress_legend = FALSE,
  add_item_title = FALSE
)
}
\arguments{
\item{dsn}{A data.frame object with PRO-CTCAE data}

\item{id_var}{A character string.Name of ID variable differentiating each
PRO-CTCAE survey/participant entered as a quoted string.}

\item{cycle_var}{A character string. Name of variable differentiating one
longitudinal/repeated PRO-CTCAE survey from another, within an individual
ID.}

\item{baseline_val}{A number indicating the expected baseline cycle/time
point.}

\item{arm_var}{A character string. Name of arm variable differentiating
treatment groups. Must be character or factor class. Overall AUC
will be reported if no arm/grouping variable is provided. Defaults to
\code{NA}.}

\item{plot_limit}{A number. Limit the number of cycles to be plotted up to
and including a given cycle number. All available cycle time points are
plotted if no cycle number is provided. Defaults to \code{NA}.}

\item{colors}{A number. Specify the coloring scheme of symptom scores within
frequency bars. Options include: 1 = Blue and red color shading, 2 =
qualitative color shades (color blind friendly), 3 = black and white.
Defaults to 1.}

\item{bar_label}{A number. Label frequency bars with sample size (n) or percent
shown on the y-axis. Label options include: \code{1} = sample size (n)
within each cycle (symptom score 0 or higher), \code{2} = sample size (n)
within each cycle with present symptoms (symptom score > 0), \code{3} =
sample size (n) within each cycle with severe symptoms (symptom score >=
3), \code{4} = percent of subjects within each cycle with present
symptoms (symptom score > 0), \code{5} = percent of subjects within each
cycle with severe symptoms (symptom score >= 3). No labels will be applied
if not specified. Defaults to \code{NA}.}

\item{cycle_label}{Logical. Assign custom labels to cycles/time point. If
\code{TRUE}, the \code{cycle_vals} and \code{cycle_labs} must also be specified.}

\item{cycle_vals}{Numeric column vector. Vector of values seen within the
\code{cycle_var} variable. Must be same length of \code{cycle_labs}. Defaults
to \code{NA}.}

\item{cycle_labs}{Character column vector. Vector of labels to be mapped to
the associated \code{cycle_vals}. Must be same length of \code{cycle_vals}.
Defaults to \code{NA}.}

\item{summary_only}{Logical. Only display the summary measures in figures /
Suppress the individual time points from plotting. Defaults to
\code{FALSE}.}

\item{summary_highlight}{Logical. Add black box around summary measure bar
chart. Defaults to \code{FALSE}.}

\item{cycles_only}{Logical. Only display the longitudinal time points in
figures / Suppress the summary measures from plotting. Defaults to
\code{TRUE}.}

\item{x_lab_angle}{A integer between 0 and 360. Allows the user to rotate the
x axis labels in order to fit long arm names (0 or 45 recommended).
Defaults to \code{0}.}

\item{x_lab_vjust}{A number. A ggplot2 object option. Allows the user to
vertically adjusts the x axis labels in order to fit arm names. Defaults to
\code{1}.}

\item{x_lab_hjust}{A number. A ggplot2 object option. Allows the user to
horizontally adjusts the x axis labels in order to fit arm names. Defaults
to \code{0}.}

\item{x_label}{A character string. Label for the x axis of the plot. Defaults
to \code{"Randomized Treatment Assignment"} if \code{arm_var} is specified,
defaults to \code{"Overall"} if not \code{arm_var} is specified.}

\item{y_label}{A character string. Label for the y axis of the plot. Defaults
to \code{"Percent of Total Frequency"}.}

\item{footnote_break}{Logical. Add a line break to the footnote Defaults to
\code{FALSE}.}

\item{suppress_legend}{Logical. Suppress the legend from appearing in figure.
Defaults to \code{FALSE}.}

\item{add_item_title}{Logical. Add the symptom item name as a title to the
figure. Defaults to \code{FALSE}.}
}
\value{
A list object. The returned object is a (k X 2) or (k x 3) nested
  list. Where k is the number of PRO-CTCAE item groups (e.g. pain, fatigue,
  nausea); list[[1 ... i ... k]]. For each list item there are 2 or 3
  elements. The 1st element of each list item is the name of the PRO-CTCAE
  item group returned as a string. The 2nd element is the PRO-CTCAE figure as
  a ggplot object. These objects can be modified as such.
}
\description{
Data format should be in 'long' format, where each PRO-CTCAE item is a
	  variable/column.
}
\examples{
\dontrun{
fig_acute = toxFigures(dsn = ProAE::tox_acute,
 cycle_var = "Cycle",
 baseline_val = 1,
 arm_var = "arm",
 id_var = "id",
 x_lab_angle = -45,
 x_lab_vjust =  .3,
 x_lab_hjust = .2,
 colors = 2)
fig_acute[[1]]
}
}
